+++ rpi_rdm_responder/lib/monitor.c
@@ -34,8 +34,8 @@
 #include "rdm_handle_data.h"
 #include "dmx_devices.h"
 
-static uint16_t function_count_previous = 0;			///<
-static uint16_t dmx_available_count_previous = 0;		///<
+static uint32_t function_count_previous = 0;			///<
+static uint32_t dmx_available_count_previous = 0;		///<
 
 void monitor_update(void)
 {
@@ -86,10 +86,10 @@ void monitor_update(void)
 	}
 
 	const struct _dmx_devices_statistics *dmx_handle_data_statistics = dmx_devices_get_statistics();
-	const uint16_t function_count_per_second = dmx_handle_data_statistics->function_count - function_count_previous;
-	const uint16_t dmx_available_count_per_second = dmx_handle_data_statistics->dmx_available_count - dmx_available_count_previous;
+	const uint32_t function_count_per_second = dmx_handle_data_statistics->function_count - function_count_previous;
+	const uint32_t dmx_available_count_per_second = dmx_handle_data_statistics->dmx_available_count - dmx_available_count_previous;
 
-	monitor_line(MONITOR_LINE_STATS, "%d / %d", function_count_per_second, dmx_available_count_per_second);
+	monitor_line(MONITOR_LINE_STATS, "%ld / %ld", function_count_per_second, dmx_available_count_per_second);
 
 	function_count_previous = dmx_handle_data_statistics->function_count;
 	dmx_available_count_previous = dmx_handle_data_statistics->dmx_available_count;
